Update ft847.c
Used udev to create a link to ttyUSB0, ttyUSBFT847.  The ttyUSB device would occasionally reset and come up as ttyUSB1.  ttyUSBFT847 still pointed to the device (thanks to udev) but it was necessary to detect the change, close the tty port and open it again.  This was done in ft847_write() function.